Scripts05/Migration/requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pandas
numpy

# Database Connectors
snowflake-connector-python
psycopg2-binary
sqlalchemy

# Data Processing
pyarrow

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
coverage==7.3.2

# Code Quality
flake8==6.1.0
black==23.12.1
isort==5.13.2

# AWS (Future)
# boto3==1.34.0

